tests/wpt/web-platform-tests/css/selectors/focus-visible-009.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <title>CSS Test (Selectors): Keyboard focus enables :focus-visible</title>
  <link rel="author" title="Alice Boxhall" href="aboxhall@chromium.org" />
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <style>
    :focus-visible {
      outline: darkgreen auto 5px;
    }

    #button:focus:not(:focus-visible) {
      background-color: tomato;;
      outline: 0;
    }
  </style>
</head>
<body>
  This test checks that any element focused via an <code>autofocus</code> attribute will have <code>:focus-visible</code> matching enabled.
  <ul id="instructions">
    <li>If the button that says "I will be focused automatically" has a red background, then the test result is FAILURE. If it has a green outline, then the test result is SUCCESS.</li>
  </ul>
  <br />
  <button id="button" autofocus tabindex="-1">I will be focused automatically.</button>
  <script>
    async_test(function(t) {
      button.addEventListener("focus", t.step_func(function() {
        assert_equals(getComputedStyle(button).outlineColor, "rgb(0, 100, 0)");
        t.done();
      }));

      // Handle the case where the button is focused before the test runs.
     if (document.activeElement === button) {
        assert_equals(getComputedStyle(button).outlineColor, "rgb(0, 100, 0)");
        t.done();
     }

    }, "Autofocus should match :focus-visible");
  </script>
</body>
</html>
